[step2-2] extract NextStepCalculator, SymbolStack clas
