+++ src/test/java/step2/NextStepCalculatorTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import step2.calculator.Expression;
+import step2.calculator.NextStepCalculator;
 import step2.calculator.Operation;
 
 import java.util.*;
@@ -33,69 +34,5 @@ class NextStepCalculatorTest {
 
         // then
         assertThat(result).isEqualTo(expected);
-    }
-
-    public static class NextStepCalculator {
-        // 모든 연산은 x, y, operation symbol 이 필요함. ex) x + y
-        private static final int OPERABLE_MINIMUM_SIZE = 3;
-
-        public int calculate(final Expression ex) {
-            final SymbolStack symbolStack = new SymbolStack(ex.getSymbols());
-
-            while (isOperable(symbolStack)) {
-                final int x = symbolStack.popToInt();
-                final Operation op = symbolStack.popToOperation();
-                final int y = symbolStack.popToInt();
-
-                final int result = op.apply(x, y);
-                
-                saveResult(symbolStack, result);
-            }
-
-            return symbolStack.popToInt();
-        }
-
-        private void saveResult(final SymbolStack symbolStack, final int result) {
-            // 중간 결과를 다시 스택에 저장
-            symbolStack.add(result);
-        }
-
-        private boolean isOperable(final SymbolStack symbolStack) {
-            return symbolStack.size() >= OPERABLE_MINIMUM_SIZE;
-        }
-    }
-
-    public static class SymbolStack {
-        final Stack<String> stack;
-
-        public SymbolStack(final List<String> symbols) {
-            this.stack = toStack(symbols);
-        }
-
-        private Stack<String> toStack(final List<String> symbols) {
-            final Iterator<String> reverseIterator = new LinkedList<>(symbols)
-                    .descendingIterator();
-
-            return StreamSupport.stream(
-                    Spliterators.spliteratorUnknownSize(reverseIterator, Spliterator.ORDERED),
-                    false)
-                    .collect(Collectors.toCollection(Stack::new));
-        }
-
-        public Operation popToOperation() {
-            return Operation.symbolOf(stack.pop());
-        }
-
-        public Integer popToInt() {
-            return Integer.parseInt(stack.pop());
-        }
-
-        public void add(final Integer number) {
-            stack.add(String.valueOf(number));
-        }
-
-        public int size() {
-            return stack.size();
-        }
     }
 }
